// filmosearch.frontend/src/api/api.ts
export interface ActorFilm {
    actorId?: number;
    actor?: Actor;
    filmId?: number;
    film?: Film;
}
// ...
export interface CreateReviewDto {
    title?: string | undefined;
    description?: string | undefined;
    stars?: number;
    filmId?: number;
}

export interface DeleteFilmFromActorDto {
    actorFirstName?: string | undefined;
    actorLastName?: string | undefined;
    filmTitle?: string | undefined;
}

export interface Film {
    id?: number;
    title?: string | undefined;
    actorFilms?: ActorFilm[] | undefined;
    reviews?: Review[] | undefined;
}
// ...
export interface Review {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    stars?: number;
    filmId?: number;
    film?: Film;
}

export interface ReviewDetailsVm {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    stars?: number;
    film?: string | undefined;
}

export interface ReviewListVm {
    reviews?: ReviewLookupDto[] | undefined;
}

export interface ReviewLookupDto {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    stars?: number;
    film?: Film;
}

export interface ReviewsLookupDto {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    stars?: number;
}

export interface UpdateActorDto {
    id?: number;
    firstName?: string | undefined;
    lastName?: string | undefined;
}

export interface UpdateFilmDto {
    id?: number;
    title?: string | undefined;
}

export interface UpdateReviewDto {
    id?: number;
    title?: string | undefined;
    description?: string | undefined;
    stars?: number;
}